# Improve performance of generic dot products