@andreasnoack 2018-06-20
Thanks for detecting this. However, this seems to have happened as part of the transition to the new iteration protocol. The version in Julia 0.6 doesn't have the type instability issue so I think we might want to understand why this happened.

cc: @Keno @stevengj