@haampie 2018-06-20
W.r.t. `dot1`, the allocations are due to:

```julia
# in dot1
if (ix == nothing) || (iy == nothing)
    break
end
```

It turns out to be necessary to write

```julia
# in dot1
ix === nothing && break
iy === nothing && break
```

It's not enough to just replace `==` by `===`, which strikes me as odd.

The `dot2` just needs that nothing-check on the second iterator, probably to convince the compiler that it is not `nothing` altough we claim to know. The other thing is that it is apparently faster to write the check again as two split statements, which makes the whole thing similar to `dot1`:

```julia
# in dot2
while true
    ix === nothing && break
    iy === nothing && break
    s += dot(ix[1], iy[1])
    ix = iterate(x, ix[2])
    iy = iterate(y, iy[2])
end
```

With these changes things are faster again:

```julia
julia> test(Float64, 100)
"Dot product using BLAS."
  35.675 ns (0 allocations: 0 bytes)
"New generic dot product."
  275.774 ns (0 allocations: 0 bytes)
"Generic dot product in LinearAlgebra/src/generic.jl."
  216.415 ns (0 allocations: 0 bytes)
"Dot product for `AbstractVector`s in LinearAlgebra/src/generic.jl."
  202.356 ns (0 allocations: 0 bytes)
"Dot product for `AbstractArray`s in LinearAlgebra/src/generic.jl."
  109.182 ns (0 allocations: 0 bytes)
```