@ranocha 2018-06-25
I don't think the test failures on appveyor and circleci are influenced by this PR. Should I make any additional changes in order to allow merging of this PR?